# visus-front/packages/api/openapi.json
{
    "openapi": "3.1.0",
    "info": { "title": "OpenAPI definition", "version": "v0" },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/estudantes/{id}": {
            "get": {
                "tags": ["Módulo de Estudantes"],
                "summary": "Obtém um Estudante de acordo com seu ID",
                "operationId": "buscarPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estudante obtido com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EstudanteSummaryDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estudante não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EstudanteSummaryDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["Módulo de Estudantes"],
                "summary": "Atualiza as informações de um Estudante",
                "operationId": "atualizar",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstudanteUpdateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Estudante atualizado com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EstudanteResponseDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro na validação dos dados enviados",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EstudanteResponseDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estudante não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EstudanteResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Módulo de Estudantes"],
                "summary": "Remove um Estudante do sistema",
                "operationId": "remover",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "responses": {
                    "204": { "description": "Estudante removido com sucesso" },
                    "404": { "description": "Estudante não encontrado" }
                }
            }
        },
        "/educadores/{id}": {
            "get": {
                "tags": ["👨‍🏫 Educadores"],
                "summary": "Buscar Educador por ID",
                "operationId": "getEducador",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EducadorResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": ["👨‍🏫 Educadores"],
                "summary": "Atualiza um Educador",
                "operationId": "updateEducador",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateEducadorDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Atualização realizada com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EducadorSummaryDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Educador não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EducadorSummaryDTO"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Dados conflitantes ou inválidos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EducadorSummaryDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/respostas": {
            "post": {
                "tags": ["Respostas"],
                "summary": "Cria uma nova resposta de texto para uma tarefa",
                "operationId": "enviarResposta",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRespostaDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/respostas/upload": {
            "post": {
                "tags": ["Respostas"],
                "summary": "Cria uma nova resposta com upload de arquivo",
                "operationId": "criarRespostaComUpload",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRespostaMultipartDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/educadores/{id}/formacoes": {
            "get": {
                "tags": ["👨‍🏫 Educadores"],
                "operationId": "listarFormacoes",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/FormacaoListResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["👨‍🏫 Educadores"],
                "summary": "Adicionar nova formação ao educador",
                "operationId": "adicionarFormacao",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FormacaoRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Formação adicionada com sucesso",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/EducadorResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/respostas/{id}/avaliar": {
            "patch": {
                "tags": ["Respostas"],
                "summary": "Avalia uma resposta enviada por um estudante",
                "operationId": "avaliarResposta",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AvaliarRespostaDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/educadores/{id}/formacoes/{formacaoId}": {
            "delete": {
                "tags": ["👨‍🏫 Educadores"],
                "summary": "Remover formação do educador",
                "operationId": "removerFormacao",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    },
                    {
                        "name": "formacaoId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "responses": {
                    "204": { "description": "Formação removida com sucesso" }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EducadorSummaryDTO": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer", "format": "int64" },
                    "nome": { "type": "string" },
                    "email": { "type": "string" },
                    "matricula": { "type": "string" },
                    "titulo": { "type": "string" }
                }
            },
...
            "Endereco": {
                "type": "object",
                "properties": {
                    "logradouro": { "type": "string", "minLength": 1 },
                    "numero": { "type": "string", "minLength": 1 },
                    "bairro": { "type": "string", "minLength": 1 },
                    "cidade": { "type": "string", "minLength": 1 },
                    "estado": { "type": "string", "minLength": 1 },
                    "cep": { "type": "string", "minLength": 1 }
                },
                "required": [
                    "bairro",
                    "cep",
                    "cidade",
                    "estado",
                    "logradouro",
                    "numero"
                ]
            },
...
            "InstituicaoDetailsDTO": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer", "format": "int64" },
                    "nome": { "type": "string" },
                    "telefone": { "type": "string" },
                    "email": { "type": "string" },
                    "cnpj": { "type": "string" },
                    "endereco": { "$ref": "#/components/schemas/Endereco" }
                }
            },
...
            "EstudanteUpdateDTO": {
                "type": "object",
                "properties": {
                    "nome": { "type": "string" },
                    "telefone": { "type": "string" },
                    "genero": {
                        "type": "string",
                        "enum": [
                            "MASCULINO",
                            "FEMININO",
                            "OUTRO",
                            "NAO_INFORMAR"
                        ]
                    },
                    "dataNascimento": { "type": "string", "format": "date" }
                }
            },
            "EstudanteResponseDTO": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer", "format": "int64" },
                    "nome": { "type": "string" },
                    "cpf": { "type": "string" },
                    "email": { "type": "string" },
                    "matricula": { "type": "string" },
                    "periodo": {
                        "type": "string",
                        "enum": ["MATUTINO", "VESPERTINO", "NOTURNO"]
                    },
                    "anoEscolar": {
                        "type": "string",
                        "enum": [
                            "PRIMEIRO_ANO",
                            "SEGUNDO_ANO",
                            "TERCEIRO_ANO",
                            "QUARTO_ANO",
                            "QUINTO_ANO",
                            "SEXTO_ANO",
                            "SETIMO_ANO",
                            "OITAVO_ANO",
                            "NONO_ANO",
                            "PRIMEIO_ANO_EM",
                            "SEGUNDO_ANO_EM",
                            "TERCEIRO_ANO_EM"
                        ]
                    },
                    "telefone": { "type": "string" },
                    "genero": {
                        "type": "string",
                        "enum": [
                            "MASCULINO",
                            "FEMININO",
                            "OUTRO",
                            "NAO_INFORMAR"
                        ]
                    },
                    "dataNascimento": { "type": "string", "format": "date" },
                    "instituicao": {
                        "$ref": "#/components/schemas/InstituicaoDetailsDTO"
                    }
                }
            },
            "UpdateEducadorDTO": {
                "type": "object",
                "properties": {
                    "nome": { "type": "string" },
                    "telefone": { "type": "string" },
                    "genero": {
                        "type": "string",
                        "enum": [
                            "MASCULINO",
                            "FEMININO",
                            "OUTRO",
                            "NAO_INFORMAR"
                        ]
                    },
                    "dataNascimento": { "type": "string", "format": "date" },
                    "matricula": { "type": "string" },
                    "titulo": { "type": "string" }
                }
            },
...
            "CreateRespostaDTO": {
                "type": "object",
                "properties": {
                    "tarefaId": { "type": "integer", "format": "int64" },
                    "estudanteId": { "type": "integer", "format": "int64" },
                    "caminhoAnexo": { "type": "string" },
                    "conteudoTexto": { "type": "string" }
                },
                "required": ["estudanteId", "tarefaId"]
            },
            "ArquivoDTO": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer", "format": "int64" },
                    "nomeOriginal": { "type": "string" },
                    "caminho": { "type": "string" },
                    "dataEnvio": { "type": "string", "format": "date-time" },
                    "tipoMime": { "type": "string" },
                    "enviadoPor": { "type": "string" }
                }
            },
            "RespostaDTO": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer", "format": "int64" },
                    "estudanteNome": { "type": "string" },
                    "nota": { "type": "number", "format": "double" },
                    "comentarioEducador": { "type": "string" },
                    "arquivo": { "$ref": "#/components/schemas/ArquivoDTO" },
                    "conteudoTexto": { "type": "string" },
                    "dataUpload": { "type": "string", "format": "date-time" }
                }
            },
            "CreateRespostaMultipartDTO": {
                "type": "object",
                "properties": {
                    "tarefaId": { "type": "integer", "format": "int64" },
                    "conteudoTexto": { "type": "string" },
                    "arquivo": { "type": "string", "format": "binary" }
                }
            },
...
            "FormacaoRequestDTO": {
                "type": "object",
                "properties": {
                    "titulo": { "type": "string" },
                    "instituicao": { "type": "string" },
                    "dataInicio": { "type": "string", "format": "date" },
                    "dataConclusao": { "type": "string", "format": "date" },
                    "descricao": { "type": "string" }
                }
            },
            "EducadorResponseDTO": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer", "format": "int64" },
                    "nome": { "type": "string" },
                    "email": { "type": "string" },
                    "matricula": { "type": "string" },
                    "titulo": { "type": "string" },
                    "formacaoAcademica": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FormacaoResponseDTO"
                        }
                    }
                }
            },
            "FormacaoResponseDTO": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer", "format": "int64" },
                    "titulo": { "type": "string" },
                    "instituicao": { "type": "string" },
                    "dataInicio": { "type": "string", "format": "date" },
                    "dataConclusao": { "type": "string", "format": "date" },
                    "descricao": { "type": "string" }
                }
            },
...
            "EstudanteSummaryDTO": {
                "type": "object",
                "properties": {
                    "id": { "type": "integer", "format": "int64" },
                    "nome": { "type": "string" },
                    "cpf": { "type": "string" },
                    "email": { "type": "string" },
                    "matricula": { "type": "string" },
                    "periodo": {
                        "type": "string",
                        "enum": ["MATUTINO", "VESPERTINO", "NOTURNO"]
                    },
                    "anoEscolar": {
                        "type": "string",
                        "enum": [
                            "PRIMEIRO_ANO",
                            "SEGUNDO_ANO",
                            "TERCEIRO_ANO",
                            "QUARTO_ANO",
                            "QUINTO_ANO",
                            "SEXTO_ANO",
                            "SETIMO_ANO",
                            "OITAVO_ANO",
                            "NONO_ANO",
                            "PRIMEIO_ANO_EM",
                            "SEGUNDO_ANO_EM",
                            "TERCEIRO_ANO_EM"
                        ]
                    },
                    "telefone": { "type": "string" },
                    "genero": {
                        "type": "string",
                        "enum": [
                            "MASCULINO",
                            "FEMININO",
                            "OUTRO",
                            "NAO_INFORMAR"
                        ]
                    },
                    "dataNascimento": { "type": "string", "format": "date" }
                }
            },
...
            "AvaliarRespostaDTO": {
                "type": "object",
                "properties": {
                    "respostaId": { "type": "integer", "format": "int64" },
                    "nota": { "type": "number", "format": "double" },
                    "comentario": { "type": "string" }
                },
                "required": ["nota", "respostaId"]
            },
...
            "FormacaoListResponseDTO": {
                "type": "object",
                "properties": {
                    "formacoes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FormacaoResponseDTO"
                        }
                    },
                    "totalRegistros": { "type": "integer", "format": "int32" }
                }
            }
        }
    }
}